pyproject.toml
[tool.poetry]
name = "bender"
version = "0.7.4"
description = "CLI admin interface for Atlassian apps."
authors = ["Charles Moreland <morelanc@ohsu.edu>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.6.8"
click = "^7.1.1"
requests = "^2.23.0"
pyyaml = "^5.3.1"
colorama = "^0.4.3"
requests_toolbelt = "^0.9.1"
mkautodoc = "^0.1.0"
jsonpath-ng = "^1.5.1"
doit = "^0.32.0"

[tool.poetry.dev-dependencies]
pytest = "^5.4.1"
mkdocs = "^1.1"

[tool.poetry.scripts]
bender = 'bender.cli:cli'
bendjira = 'bender.jira.cli:cli'
bendconfluence = 'bender.confluence.cli:cli'
bendcrowd = 'bender.crowd.cli:cli'

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
